# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sessions": {
      "post": {
        "description": "Autentica um usuário e retorna um token JWT",
        "operationId": "AuthenticateController_handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthenticateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticateResponse"
                }
              }
            }
          },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Credenciais inválidas" }
        },
        "summary": "Autenticar usuário",
        "tags": ["Authentication"]
      }
    },
    "/sessions/logout": {
      "post": {
        "description": "Realiza logout do usuário (token deve ser removido no cliente)",
        "operationId": "LogoutController_handle",
        "parameters": [],
        "responses": {
          "204": { "description": "Logout realizado com sucesso" },
          "401": { "description": "Token inválido ou não fornecido" }
        },
        "summary": "Logout do usuário",
        "tags": ["Authentication"]
      }
    },
    "/profile": {
      "get": {
        "description": "Retorna as informações do perfil do usuário autenticado",
        "operationId": "GetUserProfileController_handle",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil do usuário retornado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": { "description": "Token JWT inválido ou expirado" },
          "404": { "description": "Usuário não encontrado" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Obter perfil do usuário",
        "tags": ["User Profile"]
      }
    },
...
    "/users/{id}/profile": {
      "patch": {
        "description": "Edita o perfil do próprio usuário. Apenas o próprio usuário pode editar seu perfil.",
        "operationId": "EditUserProfileController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do usuário cujo perfil será editado",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Dados do perfil a serem atualizados",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "João Silva",
                    "description": "Nome completo do usuário"
                  },
                  "position": {
                    "type": "string",
                    "example": "Técnico em Informática",
                    "description": "Cargo do usuário"
                  }
                },
                "required": ["name", "position"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Profile updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão para editar este perfil"
          },
          "404": { "description": "Usuário não encontrado" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Editar perfil do usuário",
        "tags": ["User Profile"]
      }
    },
...
    "/categories/{id}/restore": {
      "patch": {
        "description": "Restaura uma categoria que estava na lixeira (requer role MANAGER+)",
        "operationId": "RestoreCategoryController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da categoria a ser restaurada",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": { "description": "Categoria restaurada com sucesso" },
          "400": { "description": "Categoria não encontrada" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          },
          "404": { "description": "Categoria não encontrada" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Restaurar categoria da lixeira",
        "tags": ["Categories"]
      }
    },
...
    "/locations/{id}/trash": {
      "patch": {
        "description": "Move uma localização para a lixeira (soft delete, requer role MANAGER+)",
        "operationId": "TrashLocationController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da localização a ser movida para lixeira",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Localização movida para lixeira com sucesso"
          },
          "400": { "description": "Localização não encontrada" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          },
          "404": { "description": "Localização não encontrada" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Mover localização para lixeira",
        "tags": ["Locations"]
      }
    },
    "/locations/{id}/restore": {
      "patch": {
        "description": "Restaura uma localização que estava na lixeira (requer role MANAGER+)",
        "operationId": "RestoreLocationController_handle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da localização a ser restaurada",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": { "description": "Localização restaurada com sucesso" },
          "400": { "description": "Localização não encontrada" },
          "401": { "description": "Token JWT inválido ou expirado" },
          "403": {
            "description": "Usuário não tem permissão (requer MANAGER+)"
          },
          "404": { "description": "Localização não encontrada" }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Restaurar localização da lixeira",
        "tags": ["Locations"]
      }
    },
...
  },
  "info": {
    "title": "IFAL Arapiraca Backend API",
    "description": "API para gerenciamento de problemas de infraestrutura do IFAL Arapiraca",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "AuthenticateRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "joao.silva@ifal.edu.br"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário",
            "example": "123456789"
          }
        },
        "required": ["email", "password"]
      },
      "AuthenticateResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Token JWT de acesso",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": ["access_token"]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do usuário",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Nome do usuário",
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "joao.silva@ifal.edu.br"
          },
          "position": {
            "type": "string",
            "description": "Cargo/posição do usuário",
            "example": "Técnico em Informática"
          },
          "role": {
            "type": "string",
            "description": "Role do usuário",
            "enum": ["REPORTER", "MANAGER", "DIRECTOR", "ADMIN"],
            "example": "REPORTER"
          }
        },
        "required": ["id", "name", "email", "position", "role"]
      },
...
    }
  }
}